Use valid Java package names
Updated to use valid Java packages names. This isn't a problem on older versions of Java but newer versions of Java validate this and cause build failures.
