+++ android/rustls-platform-verifier/src/main/AndroidManifest.xml
@@ -1,5 +1,3 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.onepassword.rustls_platform_verifier">
-
-</manifest>
+    package="rustls.platformverifier.android"/>
